Updated style wrt font and color
First of all, there are now serializable versions of Font and Color.
Colors were serialized in an unreadable way, fonts could not be
serialized at all.

Second, font is split in a pressed and unpressed font like all other
style elements.

Cloning is made abstract for ElementStyle, because the cloning needs
to happen in deeper levels now anyway.