Support for multiple keycodes per key
If there are multiple keys that contain the same keycodes, all of them
are lit up. The only exception if keys only have one keycode, then they
are only lit up when they are the only key with that combination in
them.

Multiple keys is only supported for keyboard keys. For other types of
keys, this will lead to... undefind behaviour.

This handles issue #20.