Version Check
Adds a version check that runs when started. This runs asynchronously
from the main thread, and it does not introduce a popup, therefore it
can safely run every time the program starts. When a new version is
found, a menu entry is added to the main menu that notifies of the new
version, and cliking it sends the user to the releases page.