Implement item selection
Selecting an element is done by clicking it. Clicking and manipulating an
element also selects it after the mouse is released. Once an element is
selected, only it can be modified. An element can be unselected by
either pressing ESC, or clicking anywhere on the background (where there
is no element).

This is in preparation of allowing manipulations with the keyboard. A
selected element also stores  the manipulation type, which allows the
manipulations by keyboard to happen exactly the same as with the mouse.

Moving over a selected element with the mouse previews new
manipulations, but  does not apply them, until the mouse is pressed.