Fix DialogResult mess
Remove DialogResult property from all buttons. I want to set it manually
in the event handlers for the property. I never noticed this before
because I only had Accept and Cancel buttons. But simply clicking those
buttons closed the dialog. This is not intended in a form where there
are other buttons.
Additionally, when clicking Ok and Cancel button, I removed the manual
closing, as setting the DialogResult has the same effect. Also made sure
the DialogResult is properly set for all close and cancel buttons.