Fix elements disappearing during properties form
The equality was not properly defined on element definitions. This
caused the code to be unable to find the element to replace in the list.
Before that, the code would replace the element after the one that was
selected, due to incorrect replace code. After replacing that, the
equality problem surfaced.