Handle style override disabling
Disabling the style override for an element would remove it from the
list of styles, but then add it back in. This caused a crash when
opening the keyboard style form after enabling and disabling an element
style. Now, the style is properly removed from the list when the
override checkbox is disabled.