More changes required in #73
Added some spotted missing information as well as documenting the Properties and Style forms. Changed the position of a paragraph so the introduction to the Properties and Style are smoother.

If the formatting or language is not in the way you like it, you can tell me and I make the changes.