Implement more advanced change tracking
Checks whether there are changes to the definition or the style, and if
either of those are saved, those changes are set back to false. This
allows a popup asking whether the user really wants to load a new
keyboard if there are any changes, because they will then be irrevocably
reverted.